+++ xknx/devices/travelcalculator.py
@@ -144,29 +144,25 @@ class TravelCalculator:
         if position_reached_or_exceeded(relative_position):
             return self._travel_to_position
 
-        travel_time = self._calculate_travel_time(relative_position)
-        if time.time() > self._last_known_postion_timestamp + travel_time:
+        remaining_travel_time = self.calculate_travel_time(
+            from_position=self._last_known_position,
+            to_position=self._travel_to_position,
+        )
+        if time.time() > self._last_known_postion_timestamp + remaining_travel_time:
             return self._travel_to_position
 
-        progress = (time.time() - self._last_known_postion_timestamp) / travel_time
-        position = self._last_known_position + relative_position * progress
-        return int(position)
+        progress = (
+            time.time() - self._last_known_postion_timestamp
+        ) / remaining_travel_time
+        return int(self._last_known_position + relative_position * progress)
 
-    def _calculate_travel_time(self, relative_position: int) -> float:
-        """Calculate time to travel to relative position."""
-        travel_direction = (
-            TravelStatus.DIRECTION_UP
-            if relative_position < 0
-            else TravelStatus.DIRECTION_DOWN
-        )
+    def calculate_travel_time(self, from_position: int, to_position: int) -> float:
+        """Calculate time to travel from one position to another."""
+        travel_range = to_position - from_position
         travel_time_full = (
-            self.travel_time_up
-            if travel_direction == TravelStatus.DIRECTION_UP
-            else self.travel_time_down
+            self.travel_time_down if travel_range > 0 else self.travel_time_up
         )
-        travel_range = self.position_closed - self.position_open
-
-        return travel_time_full * abs(relative_position) / travel_range
+        return travel_time_full * abs(travel_range) / self.position_closed
 
     def __eq__(self, other: object | None) -> bool:
         """Equal operator."""